add spring.xml

+++ src/main/java/ua/epam/spring/hometask/service/AbstractDomainObjectService.java
@@ -38,7 +38,7 @@ public interface AbstractDomainObjectService<T extends DomainObject> {
      *            id of the object
      * @return Found object or <code>null</code>
      */
-    public T getById(@Nonnull Long id);
+    public T getById(@Nonnull Integer id);
 
     /**
      * Getting all objects from storage
